requirements.txt
torch
opencv-python
numpy
tqdm
pyyaml
matplotlib